feat(core): export tryConvertPlutusMapToUtf8Record from Cardano.util
hoist from NftMetadata.fromPlutusData and add tests
