+++ packages/core/src/Cardano/util/index.ts
@@ -5,4 +5,4 @@ export * from './txSubmissionErrors';
 export * from './resolveInputValue';
 export * from './phase2Validation';
 export * from './addressesShareAnyKey';
-export * from './isPlutusData';
+export * from './plutusDataUtils';
